backend/requirements.txt
chatterbox-tts
fastapi
uvicorn[standard]
python-multipart
torchaudio
soundfile
transformers<4.47
